Add store functionality
The store buttons work and are adding items in the cart dynamically. The remove buttons also work and are updating the cart dynamically. Finally the purchase button displays an alert when clicked to alert the customer that the purchase has been made and then deletes all the items in the cart.